Choice

Description

Documentation

class Choice a where

Class for types that allow a value to be symbolically selected from a pair of alternatives based on the value of a Bit.

This is a generalization of the choose function on Bit

Methods

choice

Arguments

:: a
false branch

-> a
true branch

-> Bit
selector

-> a