Count

Description

This type allows us to symbolically count numbers incrementing and decrementing by one. Counts can be ordered and compared for equality.

Use encode to make constant valued counts.

Use increment, decrement, and addBit to update Count values.

>>> :{
solve $
  do (x,y,z) <- exists
     let count = foldr addBit (encode 0)
     assert (count [x,y,z] === encode 2)
     assert (count [x,  z] === encode 1)
     assert (count [  y,z] === encode 1)
     return (x,y,z)
:}
Just (True,True,False)

Documentation

data Count

This type represents a natural number and supports operations like increment and comparison.

increment :: Count -> Count

Increment the value of a Count by one.

decrement :: Count -> Count

Decrement the value of a Count by one. Decrement of zero is zero.

addBit :: Bit -> Count -> Count

Add a bit representing a zero or one to the count.