LightsOut

Description

https://en.wikipedia.org/wiki/Lights_Out_(game)

This module implements a solver for the Lights Out game. The solver works by encoding the lights out problem in terms of boolean satisfiability using the Ersatz library.

This module does not attempt to produce a solution with the minimal number of button presses. This enhancement is left as an exercise for the reader!

Documentation

data LightsOut

Specification of a Lights Out puzzle

Constructors

LightsOut

Fields

dimensions :: Dim

lightsOn :: Set Coord

exampleLightsOut :: LightsOut

Example puzzle that spells ICFP.

smallLightsOut :: LightsOut

Small 3-by-3 puzzle

unsolvableLightsOut :: LightsOut

Example of a puzzle that has no solution.

data Coord

Grid coordinate

Constructors

C Int Int
column row

neighborhood

Arguments

:: Coord
coordinate

-> [Coord]
north, west, self, east, south

Cardinal direction neighborhood for a coordinate. Results include the given coordinate.

>>> neighborhood (C 2 3)
[C 2 2,C 1 3,C 2 3,C 3 3,C 2 4]

data Dim

Board dimensions

Constructors

Dim Int Int
width height

dimCoords :: Dim -> [Coord]

Generate a list of all coordinates in a board of the given size.

>>> dimCoords (Dim 3 2)
[C 1 1,C 1 2,C 2 1,C 2 2,C 3 1,C 3 2]

existsSolution :: Dim -> Ersatz (Map Coord Bit)

Generate a map associating coordinates that can be toggled with whether or not that coordinate was toggled.

isValidSolution

Arguments

:: LightsOut
puzzle

-> Map Coord Bit
clicked coordinates

-> Bit
solution is valid

Predicate for determining if a set of toggles results in all lights being turned off. A clicked coordinate corresponds to an entry in the given map with a true value.

findSolution

Arguments

:: LightsOut
puzzle

-> IO (Maybe (Map Coord Bool))
solution if one exists

Find a solution to the given puzzle that also satisfies an additional predicate if one exists.

>>> findSolution smallLightsOut
Just (fromList [(C 1 1,True),(C 1 2,False),(C 1 3,False),(C 2 1,False),(C 2 2,False),(C 2 3,False),(C 3 1,False),(C 3 2,False),(C 3 3,True)])

lightsOut

Arguments

:: LightsOut
puzzle

-> IO ()
print solution to puzzle

Attempt to solve the given puzzle and print the solution to stdout.

>>> lightsOut smallLightsOut
*..
...
..*

parsePuzzle :: String -> LightsOut

Parse a lights out puzzle encoded as lines of . for off and * for on.

Example input

*.*.
.*..
...*

>>> parsePuzzle "*.*.\n.*..\n...*\n"
LightsOut {dimensions = Dim 4 3, lightsOn = fromList [C 1 1,C 2 2,C 3 1,C 4 3]}

renderSolution

Arguments

:: Dim
board dimensions

-> Map Coord Bool
solution map

-> String
rendered solution

Render a lights out puzzle using . and * as in parsePuzzle.