Select

Description

Symbolic selection from a finite list of alternatives. This module allows using normal Haskell values symbolically.

We can extract the selected value using either runSelect or decode.

Select values can be symbolically compared using comparisons of their concrete possibilities. For example (===) will use (==) on the underlying choices.

>>> :{
solve $
  do let letters = 'a':|"bcdefg"
     x <- selectList letters
     y <- selectList letters
     assert (pure 'c' <? x && x <? y && y <? pure 'f')
     return (x,y)
:}
Just ('d','e')

Types

data Select a

A symbolic choice of regular Haskell values.

Operations

runSelect :: Choice a => Select a -> a

Extract the symbolic value contained in the selection as long as the underlying values support symbolic choice.

selectList :: NonEmpty a -> Ersatz (Select a)

Create a selection given a list of alternatives.

selectPermutationN

Arguments

:: Int
desired permutation length

-> [a]
source elements

-> Ersatz [Select a]
selected permutation

Select an ordered permutation from a list of source elements.