TotalMap

Description

This module implements a total, ordered map. Upon construction of a new map, a default value must be specified. This value will be returned for any lookup that has not otherwise been assigned.

Having a built-in default value is particularly useful when working with symbolic values as ensures every key has a value, so the structure of the map will not depend on symbolic values.

To construct new TotalMap values, use fromList and pure.

To update TotalMap values, use (=:) and fmap.

To query TotalMap values, use (!).

To combine TotalMap values, use choice and (<*>).

Documentation

data TotalMap k v

A total map behaves like a function that can be efficiently updated. The map tracks a default value when created.

fromList

Arguments

:: Ord k

=> v
default value

-> [(k, v)]
key-value pairs

-> TotalMap k v

Create a TotalMap given a list of key-value pairs and a default value for any unassigned keys. Assignments are processed left-to-right. Later assignments overwrite earlier ones.

>>> fromList 0 [('a',1), ('b',2)]
fromList 0 [('a',1),('b',2)]

>>> fromList 0 [('a',1), ('a',2)]
fromList 0 [('a',2)]

>>> fromList 0 [('a',0)]
fromList 0 [('a',0)]

(!) :: Ord k => TotalMap k v -> k -> v

Lookup the value associated with a given key in the map.

>>> let m = fromList 0 [('a',1), ('b',2)]

>>> m ! 'b'
2

>>> m ! 'c'
0

(=:) :: Ord k => k -> v -> TotalMap k v -> TotalMap k v

Assign a new value to a given key in a map.

Note that this assignment makes no attempt to avoid duplicating the default value.

>>> let m = fromList 0 [('a',1), ('b',2)]

>>> ('b' =: 0) m
fromList 0 [('a',1),('b',0)]

>>> ('b' =: 10) m
fromList 0 [('a',1),('b',10)]

>>> ('c' =: 10) m
fromList 0 [('a',1),('b',2),('c',10)]