Zhed

Description

This module implements a solver for the Zhed game.

In Zhed the player is presented with a grid of cells that can be either marked or unmarked. Additionally one or more cells will be designated as the target cell. In the initial puzzle state each of the pre-marked cells will have a number on it. Each of these numbered cells can be activated exactly once. When activated a cardinal direction is chosen. The number indicates how many of the nearest, unmarked cells will become marked.

The puzzle is considered solved when a target cell is marked.

https://play.google.com/store/apps/details?id=com.groundcontrol.zhed&hl=en_US

For more puzzle files, see https://github.com/glguy/5puzzle/tree/master/zhed-puzzles

Documentation

data Coord

Grid coordinate

Constructors

C Int Int
column row

data Dir

Cardinal directions: up, down, left, right

Constructors

U
up

D
down

L
left

R
right

data Dim

Board dimensions

Constructors

Dim Int Int
width height

data Puzzle

Puzzle parameters: board dimensions, numbered cells, target cells

Constructors

Puzzle Dim [(Coord, Int)] [Coord]

largeZhed :: Puzzle

An 8-by-8 puzzle

smallZhed :: Puzzle

A 3-by-3 puzzle

cellsFromList

Arguments

:: [Coord]
coordinate list

-> TotalMap Coord Bit
map listed coordinates set true

Construct a mapping from coordinates to bits suitable for use as a symbolic set.

coordsList

Arguments

:: Dim
board dimensions

-> Coord
starting coordinate

-> Dir
direction

-> [Coord]
list of coordinates in that direction

Produce a list of coordinates generated starting at a given location, traveling in a given direction, until the edge of the board is reached. The starting location is not included.

Valid locations are considered to be from 1 to the height or width.

Examples

....
....
.S..

>>> let dim = Dim 4 3; s = C 2 3
>>> coordsList dim s U
[C 2 2,C 2 1]
>>> coordsList dim s D
[]
>>> coordsList dim s L
[C 1 3]
>>> coordsList dim s R
[C 3 3,C 4 3]

applyMove

Arguments

:: Dim
board dimensions

-> TotalMap Coord Bit
filled cells

-> Select (Coord, Int, Dir)
chosen square and direction

-> TotalMap Coord Bit
updated cells

Given a board's dimensions and the map of cells, update the cell map with a symbolically chosen square activation.

spreadCell

Arguments

:: Int
squares available to place

-> (Count, TotalMap Coord Bit)
squares placed and current board

-> Coord
coordinate to place on

-> (Count, TotalMap Coord Bit)
updated placement count and board

Attempt to mark a particular cell on the given board. Only mark a cell if there the count of cells placed is less than the total number available.

combineChoices :: Select (Coord, Int) -> Select Dir -> Select (Coord, Int, Dir)

Combine the choice of square to activate with a direction to get back a single choice containing both.

solutionExists

Arguments

:: Puzzle
puzzle parameters

-> Ersatz [Select (Coord, Int, Dir)]
puzzle solution

Generate a solution that satisfies the clues in the given puzzle. The solution is represented as an ordered list of the squares to activate and the order to activate them in.

findSolution :: Puzzle -> IO (Maybe [(Coord, Int, Dir)])

Run the SAT solver on the solution generated by solutionExists

>>> findSolution smallZhed
Just [(C 1 2,2,R),(C 3 1,1,D)]

zhed :: Puzzle -> IO ()

Solve a given Zhed puzzle and print the solution to stdout.

>>> zhed smallZhed
 .  .  2v
 1> .  .
 .  .  X

parsePuzzle :: String -> Puzzle

Parse a puzzle representing square by their number value, targets with an X and empty space with .

Example

..1
2..
..X

renderSolution

Arguments

:: Puzzle
puzzle

-> [(Coord, Int, Dir)]
solution order

-> String
rendered solution

Render a puzzle and its solution. Each square in the puzzle will be changed to be its sequence number in a satisfying solution and an indicator of the direction in which that square should be expanded.